+++ example.env
@@ -1,0 +1,7 @@
+REACT_APP_API_KEY=AIzaSyB8WHjVvvEXHH9-0Z5AvnN2umTQKaQ7Ovs
+REACT_APP_AUTH_DOMAIN=fir-tutorial-d893c.firebaseapp.com
+REACT_APP_DATABASE_URL=https://fir-tutorial-d893c.firebaseio.com
+REACT_APP_PROJECT_ID=fir-tutorial-d893c
+REACT_APP_STORAGE_BUCKET=fir-tutorial-d893c.appspot.com
+REACT_APP_MESSAGING_SENDER_ID=151360389556
+REACT_APP_APP_ID=1:151360389556:web:b5f69325e6a6bc12ceb2e1
