+++ src/components/App/index.js
@@ -6,14 +6,14 @@ import {
 
 import Navigation from '../Navigation';
 import LandingPage from '../Landing';
-import SignUpPage from '../SignUp';
-import SignInPage from '../SignIn';
-import PasswordForgetPage from '../PasswordForget';
+import SignUpPage from '../auth/SignUp';
+import SignInPage from '../auth/SignIn';
+import PasswordForgetPage from '../auth/PasswordForget';
 import HomePage from '../Home';
 import AccountPage from '../Account';
 import AdminPage from '../Admin';
 import * as ROUTES from '../../constants/routes';
-import { withAuthentication } from '../Session';
+import { withAuthentication } from '../auth/Session';
 
 
 const App = () => (
