+++ src/components/Navigation/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import { Link } from 'react-router-dom';
-import SignOutButton from '../SignOut';
+import SignOutButton from '../auth/SignOut';
 import * as ROUTES from '../../constants/routes';
-import { AuthUserContext } from '../Session';
+import { AuthUserContext } from '../auth/Session';
 
 const Navigation = () => (
     <div>
