+++ src/components/auth/Session/withAuthentication.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AuthUserContext from './context';
-import { withFirebase } from '../Firebase';
+import { withFirebase } from '../../Firebase';
 
 
 const withAuthentication = Component => {
