+++ src/components/auth/SignOut/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { withFirebase } from '../Firebase';
+import { withFirebase } from '../../Firebase';
 
 const SignOutButton = ({ firebase }) => ( // 受け取ったpropsからfirebaseだけ抜き取る
     <button type="button" onClick={firebase.doSignOut}>
