+++ src/components/auth/SignUp/index.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Link, withRouter } from 'react-router-dom';
 import { compose } from 'recompose';
-import * as ROUTES from '../../constants/routes';
-import { withFirebase } from '../Firebase';
+import * as ROUTES from '../../../constants/routes';
+import { withFirebase } from '../../Firebase';
 
 const SignUpPage = () => (
     <div>
